requirements.txt
fastapi
uvicorn
sqlalchemy==1.4.45
psycopg2-binary==2.9.6
pillow==10.4.0
python-multipart==0.0.9
easyocr==1.7.1